edit on home fragment

+++ data/cache/src/main/java/com/example/cache/models/RatingEntity.kt
@@ -2,5 +2,5 @@ package com.example.cache.models
 
 data class RatingEntity(
     val quantity: Int,
-    val rate: Int
+    val rate: Float
 )
